refactored radio buttons into its own component
